pyproject.toml
[project]
name = "credit-default-mlops-pipeline"
version = "0.1.0"
description = "Credit Default Scoring API (FastAPI + MLflow + Prometheus + Grafana)"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

# Minimal runtime deps for local dev; CI will typically install requirements*.txt
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "mlflow>=2.10",
  "prometheus-fastapi-instrumentator>=6.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
